requirements.txt
python-telegram-bot==20.3
firebase-admin==6.2.0
openai==0.28
python-dotenv==1.0.0
google-cloud-firestore==2.11.0